refactor(providers): share runtime stream helpers
