+++ src/ouroboros/providers/codex_cli_adapter.py
@@ -8,7 +8,7 @@ without requiring an API key.
 from __future__ import annotations
 
 import asyncio
-from collections.abc import AsyncIterator, Callable
+from collections.abc import Awaitable, Callable
 import contextlib
 import json
 import os
@@ -41,11 +41,7 @@ from ouroboros.providers.base import (
     MessageRole,
     UsageInfo,
 )
-from ouroboros.providers.codex_cli_stream import (
-    collect_stream_lines,
-    iter_stream_lines,
-    terminate_process,
-)
+from ouroboros.providers.codex_cli_stream import RuntimeStreamMixin, collect_stream_lines
 from ouroboros.providers.profiles import resolve_completion_profile_result
 from ouroboros.providers.retry import TRANSIENT_ERROR_PATTERNS
 
@@ -84,7 +80,7 @@ _CODEX_PROVIDER_MARKERS = (
 _OPENAI_RESPONSES_ENDPOINT = "api.openai.com/v1/responses"
 
 
-class CodexCliLLMAdapter:
+class CodexCliLLMAdapter(RuntimeStreamMixin):
     """LLM adapter backed by local Codex CLI execution.
 
     Streaming progress callback contract (``on_message``):
@@ -115,6 +111,12 @@ class CodexCliLLMAdapter:
     _max_ouroboros_depth = DEFAULT_MAX_OUROBOROS_DEPTH
     _child_session_env_keys = DEFAULT_CODEX_CHILD_SESSION_ENV_KEYS
     _completion_profile_backend = "codex"
+
+    def _stream_provider_name(self) -> str:
+        return "codex_cli"
+
+    def _stream_line_collector(self) -> Callable[..., Awaitable[list[str]]]:
+        return collect_stream_lines
 
     def __init__(
         self,
@@ -740,30 +742,6 @@ class CodexCliLLMAdapter:
             )
             tool_info = self._format_tool_info("WebSearch", {"query": query})
             self._on_message("tool_started" if event_type == "item.started" else "tool", tool_info)
-
-    async def _iter_stream_lines(
-        self,
-        stream: asyncio.StreamReader | None,
-        *,
-        chunk_size: int = 16384,
-    ) -> AsyncIterator[str]:
-        """Yield decoded lines without relying on StreamReader.readline()."""
-        async for line in iter_stream_lines(stream, chunk_size=chunk_size):
-            yield line
-
-    async def _collect_stream_lines(
-        self,
-        stream: asyncio.StreamReader | None,
-    ) -> list[str]:
-        """Drain a subprocess stream without blocking stdout event parsing."""
-        return await collect_stream_lines(stream)
-
-    async def _terminate_process(self, process: Any) -> None:
-        """Best-effort subprocess shutdown used for timeouts and cancellation."""
-        await terminate_process(
-            process,
-            shutdown_timeout=self._process_shutdown_timeout_seconds,
-        )
 
     def _prompt_stdin_bytes(self, prompt: str) -> bytes | None:
         """Return bytes to write to process stdin for the prompt, if any."""
